requirements.txt
Django==2.2.13
psycopg2-binary==2.8.4
django-environ==0.4.5
gunicorn==20.0.4
whitenoise==4.1.4
requests==2.22.0
python-dateutil==2.8.1
docker==4.1.0
beautifulsoup4==4.8.2
lxml==4.4.2
django-storages==1.8
boto3==1.5.0
PyGithub==1.43.8
sentry-sdk==0.14.1
vcrpy==2.0.1
django-db-geventpool==3.2.0
gevent==1.4.0
pylint==2.4.4
pylint-django==2.0.13
black==19.10b0
randomcolor==0.4.4.5
django-rangepaginator==1.2.0
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-api-key==2.0.0
django-filter==2.2.0
djangorestframework==3.10.2
